backend/requirements.txt
Django==5.0.6
djangorestframework==3.14.0
django-cors-headers==4.3.1
djangorestframework-simplejwt==5.3.1
django-environ==0.11.2
razorpay==1.4.2
Pillow==10.2.0